1. Добавлен аттрибут JsonConstructor на конструктор класса Envelope.cs 2. Добавлен проект с интеграционными тестами 3. Добавлен тест на проверку полной multi-part загрузки

+++ backend/FileService/src/FileService.Contracts/MediaAssets/StartMultiPartUpload/StartMultiPartUploadResponse.cs
@@ -4,4 +4,4 @@ public record StartMultiPartUploadResponse(
     Guid MediaAssetId,
     string UploadId,
     IReadOnlyList<ChunkUploadUrl> ChunkUploadUrls,
-    long ChunkSize);
+    int ChunkSize);
